requirements.txt
requests
pip
discord.py
random